Move routes to separate folder. Clean route props to be more uniform.
