+++ src/routes/index.ts
@@ -1,0 +1,4 @@
+export { default as Home } from './Home';
+export { default as Login } from './Login';
+export { default as Inventory } from './Inventory';
+export { default as NotFound } from './NotFound';
